+++ src/UserService.Mappers/UserMapper.cs
@@ -3,18 +3,15 @@ using LT.DigitalOffice.UserService.Models.Db;
 using LT.DigitalOffice.UserService.Models.Dto;
 using System;
 using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
 
 namespace LT.DigitalOffice.UserService.Mappers
 {
     /// <summary>
-    /// Represents mapper. Provides methods for converting an object of <see cref="DbUser"/> type into an object of <see cref="User"/> type according to some rule.
+    /// Represents mapper. Provides methods for converting an object of <see cref="DbUser"/>
+    /// type into an object of <see cref="User"/> type according to some rule.
     /// </summary>
     public class UserMapper : IMapper<DbUser, User>, IMapper<DbUser, string, object>,
-        IMapper<UserCreateRequest, DbUser>, IMapper<EditUserRequest, DbUser>,
-        IMapper<EditUserRequest, string, DbUserCredentials>,
-        IMapper<UserCreateRequest, Guid, DbUserCredentials>
+        IMapper<UserCreateRequest, DbUser>, IMapper<EditUserRequest, DbUser>
     {
         public User Map(DbUser dbUser)
         {
@@ -101,40 +98,6 @@ namespace LT.DigitalOffice.UserService.Mappers
                 AvatarFileId = null,
                 IsActive = true,
                 IsAdmin = request.IsAdmin
-            };
-        }
-
-        public DbUserCredentials Map(UserCreateRequest request, Guid userId)
-        {
-            if (request == null)
-            {
-                throw new ArgumentNullException(nameof(request));
-            }
-
-            return new DbUserCredentials
-            {
-                UserId = userId,
-                Email = request.Email,
-                PasswordHash = Encoding.UTF8.GetString(new SHA512Managed().ComputeHash(
-                    Encoding.UTF8.GetBytes(request.Password))),
-                Salt = "Exmple_salt"
-            };
-        }
-
-        public DbUserCredentials Map(EditUserRequest request, string salt)
-        {
-            if (request == null)
-            {
-                throw new ArgumentNullException(nameof(request));
-            }
-
-            return new DbUserCredentials
-            {
-                UserId = request.Id,
-                Email = request.Email,
-                PasswordHash = Encoding.UTF8.GetString(new SHA512Managed().ComputeHash(
-                    Encoding.UTF8.GetBytes(request.Password))),
-                Salt = salt
             };
         }
     }
